Move cell_* methods and tests
